STS-435: Change clean URLs not to rely on internal forwards
Must not cache the whole parameter map. During an internal forward, the underlying request object is replaced by a new one. Any parameters in the query string of the forward are then hidden because the original parameter map was cached by this class. Instead, I have added a Map implementation that merges parameter values on the fly.

git-svn-id: https://stripes.svn.sourceforge.net/svnroot/stripes/trunk@628 d471f07e-d8fc-0310-a67b-dd4c8ff0e7cd