STS-435: Change clean URLs not to rely on internal forwards.
It was flawed logic to try to maintain a stack of requests so that we always have access to any parameters that might be added by wrappers above the stripes wrapper. If a wrapper above the stripes wrapper modifies the parameters then it is up to that wrapper to deal with merging the values. The request stack is gone now.

git-svn-id: https://stripes.svn.sourceforge.net/svnroot/stripes/trunk@645 d471f07e-d8fc-0310-a67b-dd4c8ff0e7cd