The stripes filter now accepts a new init-param - Extension.Packages. It is a comma separated list of packages that will automatically be searched for ConfigurableComponents, TypeConverters, and Formatters.
ConfigurableComponents explicitly listed in web.xml override those found in the listed extension packages.

TypeConverters and Formatters will be added to their respective factories to handle the class in the type parameter unless that class is Object. Additional classes that the TypeConverter or Formatter handles may be specified via a TargetTypes annotation. TargetTypes may include classes, interfaces, and annotations. An annotation target type applies when the type being converted to or formatted is annotated with the specified annotation and a more suitable TypeConverter or Formatter cannot be found.

git-svn-id: https://stripes.svn.sourceforge.net/svnroot/stripes/trunk@734 d471f07e-d8fc-0310-a67b-dd4c8ff0e7cd