First cut at a fix for STS-360.  We now have three implementations of EL validation available, and the code tries to pick the best one that will work in the container. In many cases this will result in an implementation that no longer needs to allocate PageContexts in the dispatcher servlet.
Side effects of this include upgrading our Servlet and JSP API jars and adding the new EL API.  We'll have to be careful from now on to not accidentally use Servlet 2.3 or JSP 2.1 features in the main code.

I'm pretty sure this will need some more testing on a Servlet 2.3 container before we release it!

git-svn-id: https://stripes.svn.sourceforge.net/svnroot/stripes/trunk@791 d471f07e-d8fc-0310-a67b-dd4c8ff0e7cd