Fixed STS-620 the best I could. BootstrapPropertyResolver will catch AccessControlExceptions when getting property values and log a debug message telling what happened if it can't read the property. ResolverUtil has changed significantly since this problem was reported and does not appear to have the same problems as before. Aside from the two changes noted below, everything worked fine after this code change.
Stripes still requires java.lang.RuntimePermission "accessDeclaredMembers", which is not in the default catalina.policy. It needs that one to find the event handler methods in the ActionBean classes.

I also had to allow file read for WEB-INF/classes/logging.properties. That appears to be normal since there was already a template in the catalina.policy file for it.

git-svn-id: https://stripes.svn.sourceforge.net/svnroot/stripes/trunk@1027 d471f07e-d8fc-0310-a67b-dd4c8ff0e7cd